Quickpay: Adds currency for v7 store operations
While working with Quickpay in production mode, it has been discovered
that "currency" is a required value for store (subscribe) operations in
the v7 API.  Without it, subscribe operations complain with "Error in
request parameters". This is confirmed by their documentation, which
currently states that currency is required:
http://doc.quickpay.net/api/messagetypes.html#index2h2

This commit makes the change such that only the subscribe parameters for
v7 (and up) are changed.  New unit tests were added to prove the correct
parameters are now being sent, and all existing remote tests were run
and are passing.