[fusebox] Switch store from $0 auth to $1 auth
It appears that not all banks support the $0 auth, even though the
fusebox documentation recommends it as the way to store a card.

This switches from a $0 auth to a $1 auth in an attempt at a more
reliable way to store cards.  Note that Visa mandates warn that there
may be fees for "Authorization System Misuse", but since support for $0
seems to still be problematic, we don't feel these fees will ever
materialize.