Authorize.net CIM - Handle multiple error messages
Authorize.net can send back one or multiple errors in their responses.
We need to correctly assess whether response["messages"]["message"] is
an Array or Hash, and if it is an Array, then we need to set the top
level error_code and message fields to the first error returned.

Closes #2537