Always send currency and amount to Quickpay on store card
This removes the condition `if @protocol >= 7` so the amount and the currency will be sent to `:subscribe` requests made over the v3 protocol. In addition this updates the tests to confirm the new functionality.

The reason for this is a change on the part of Quickpay now requiring this information to be present.